+++ source/test.ts
@@ -1,14 +1,12 @@
-'use strict'
-
 // Import
-const { equal } = require('assert-helpers')
-const { suite } = require('kava')
-const { Logger } = require('caterpillar')
-const Human = require('./')
-const { PassThrough } = require('stream')
+import { equal } from 'assert-helpers'
+import { suite } from 'kava'
+import { Logger } from 'caterpillar'
+import Human from './index.js'
+import { PassThrough } from 'stream'
 
 // Prepare
-function cleanChanging(item) {
+function cleanChanging(item: string) {
 	item = item
 		.replace(/\[\d{4}-\d{2}-\d{2} \d{2}:\d{2}:\d{2}.\d{3}\]/, 'date')
 		.replace(/\[[/\\].+?:\d{1,}\]/, 'file')
@@ -38,12 +36,17 @@ suite('human', function (suite) {
 		})
 	})
 
-	function addSuite(name, config, expected, cleaner) {
+	function addSuite(
+		name: string,
+		config: { color?: boolean; level?: number },
+		expected: string[],
+		cleaner?: typeof cleanChanging
+	) {
 		suite(name, function (suite, test) {
 			const logger = new Logger()
 			const browser = new Human(config)
 			const output = new PassThrough()
-			let actual = []
+			let actual: string[] = []
 			if (cleaner) expected = expected.map(cleaner)
 
 			output.on('data', function (chunk) {